requirements.txt
discord.py==2.3.2
spotipy==2.23.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
matplotlib==3.8.2
Pillow==10.1.0
aiohttp==3.9.1
pytz==2023.3
pandas==2.1.4